feat(tensor-tree): add incident patch drawing for TT
This commit adds a new html file that shows how we can use `d3.geoProjection` to visualize tensor-tree patches. The implementation addresses both issues that are discussed here: (https://github.com/ladybug-tools/bsdf-viewer/issues/5).

The trick was to implement each patch as a d3-geo-shape which is mostly used for drawing maps! Because each patch is now a goJSON polygon it translates to a D3 native object. I attach an id to each patch to show how it is possible. Then once the patches are generated we can use d3.projection to visualize them. This makes interactions much easier as we can bind the data for each patch to the geometry.

With this approach we also keep the original orthogonal coordinates for each patch and use that for Boolean functions like merging them together. Running union calculation on the original grid of squares is quite fast. For testing I implemented it to the current interface and it works almost real time. Try to click on patches and the algorithm will join the ones that are next to each other together and draws a curve around them.

![merge patches](https://user-images.githubusercontent.com/2915573/64070739-ad577100-cc36-11e9-9347-d8678639e12d.gif)

This will be helpful for joining outgoing patches based on values.  Basically we can group all the patches with similar values regardless of being neighbors or not and then use `polygon-clipping` library to join them together. Since the patches is implemented in geoJSON format we can take advantage of all the different libraries to interact with geoJSON files including `polygon-clipping`! :)

Next step is to figure out how to map data from BSDF file to these patches. Though the parser works I still have no idea how to match the values to these patches. @andyrew is it something that you can help me with or should I ask Greg?

I tried to figure it out based on your code but I get lost when you use Radiance to get the values. It is hard to imagine what is the result structure that is generated from Radiance. I wish I could have get TensorTree code working on my machine! That would have made everything easier. :/

resolves #5